:recycle: Refactor: redis config 파일 이동
<footer>
- 관련: #142
