--- storage/db-redis/src/main/java/com/namo/spring/db/redis/Main.java
@@ -1,7 +1,0 @@
-package com.namo.spring.db.redis;
-
-public class Main {
-	public static void main(String[] args) {
-		System.out.println("Hello world!");
-	}
-}
